@@ -30,9 +30,13 @@ NOTE:  This code was generated by the GeneratePackageVersions tool. To safely
 @"  </ItemGroup>
 </Project>";
 
+        private const string TargetFrameworkFormat =
+@"
+      <TargetFramework>{0}</TargetFramework>";
+
         private const string EntryFormat =
 @"    <PackageVersionSample Include=""samples*\**\{0}.csproj"">
-      <Properties>ApiVersion={1};TargetFramework=netcoreapp2.1;RestoreRecursive=false;BuildProjectReferences=false</Properties>
+      <Properties>ApiVersion={1};RestoreRecursive=false;BuildProjectReferences=false</Properties>{2}
     </PackageVersionSample>";
 
         public MSBuildPropsFileGenerator(string filename)
@@ -56,7 +60,7 @@ NOTE:  This code was generated by the GeneratePackageVersions tool. To safely
             }
         }
 
-        public override void Write(string integrationName, string sampleProjectName, IEnumerable<string> packageVersions)
+        public override void Write(PackageVersionEntry packageVersionEntry, IEnumerable<string> packageVersions)
         {
             Debug.Assert(Started, "Cannot call Write() before calling Start()");
             Debug.Assert(!Finished, "Cannot call Write() after calling Finish()");
@@ -64,7 +68,8 @@ NOTE:  This code was generated by the GeneratePackageVersions tool. To safely
             StringBuilder bodyStringBuilder = new StringBuilder();
             foreach (string packageVersion in packageVersions)
             {
-                FileStringBuilder.AppendLine(string.Format(EntryFormat, sampleProjectName, packageVersion));
+                string targetFrameworkProperty = string.IsNullOrEmpty(packageVersionEntry.SampleTargetFramework) ? string.Empty : string.Format(TargetFrameworkFormat, packageVersionEntry.SampleTargetFramework);
+                FileStringBuilder.AppendLine(string.Format(EntryFormat, packageVersionEntry.SampleProjectName, packageVersion, targetFrameworkProperty));
             }
         }
     }
